modify some models

+++ models/DecorationCompany.php
@@ -21,6 +21,6 @@ class DecorationCompany extends ActiveRecord
      */
     public static function totalNumber()
     {
-        return self::find()->count();
+        return (int)self::find()->count();
     }
 }
