+++ models/Manager.php
@@ -21,6 +21,6 @@ class Manager extends ActiveRecord
      */
     public static function totalNumber()
     {
-        return self::find()->count();
+        return (int)self::find()->count();
     }
 }
