+++ models/Supplier.php
@@ -224,7 +224,7 @@ class Supplier extends ActiveRecord
      */
     public static function totalNumber()
     {
-        return self::find()->count();
+        return (int)self::find()->count();
     }
 
     /**
